debezium/requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.0
openai>=1.0.0
